Double to float narrowing fixed.

+++ replica.cpp
@@ -16,7 +16,7 @@ using namespace std;
 /**************************************************************
 *  Constructor 
 **************************************************************/
-Replica::Replica(unsigned short _Nx, unsigned short _Ny, float _delta, float _T, long seed, int _id):
+Replica::Replica(unsigned short _Nx, unsigned short _Ny, double _delta, float _T, long seed, int _id):
 //Initialize random objects with default values
  RandomBase(seed)
 
@@ -63,7 +63,7 @@ Replica::Replica(unsigned short _Nx, unsigned short _Ny, float _delta, float _T,
     if  (delta < 1.0) epsilon = (1.0 - delta)/4.0;
     else              epsilon = 0;
     //Set vertex weights 
-    float tmp2[6] = { epsilon            , epsilon,
+    double tmp2[6] = { epsilon            , epsilon,
                       delta/2.0 + epsilon, delta/2.0 + epsilon,
                       0.5,                 0.5}; 
     memcpy(VertexWeight, tmp2, sizeof tmp2);
